- fixed 64 bits compiling warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@6412 689a6050-402a-0410-94f2-e92a70836424

+++ parser/parse_content.c
@@ -270,7 +270,7 @@ char* decode_mime_type(char *start, char *end, unsigned int *mime_type, content_
 
 	p = start;
 
-	LM_DBG("Decoding MIME type for:[%.*s]\n",end-start,start);
+	LM_DBG("Decoding MIME type for:[%.*s]\n",(int)(end-start),start);
 
 	/* search the begining of the type */
 	while ( p<end && (*p==' ' || *p=='\t' ||
